platform/admin-api/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database
alembic==1.13.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Cloud SQL Python Connector (for IAM auth with private IP)
cloud-sql-python-connector[pg8000]==1.11.0
pg8000==1.31.2

# Google Auth (for ID token verification)
google-auth>=2.28.0
google-auth-oauthlib==1.2.0

# GCP SDK for platform operations
google-cloud-iam>=2.14.0
google-cloud-compute>=1.15.0
google-cloud-artifact-registry>=1.11.0
google-cloud-build>=3.22.0
google-cloud-resource-manager>=1.12.0
google-cloud-service-usage>=1.8.0
google-api-core>=2.15.0

# Additional utilities
python-multipart==0.0.6
requests==2.31.0
